Relocate Test folder under src
